# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bank App API ",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/client-registers": {
            "post": {
                "tags": [
                    "Client register"
                ],
                "summary": "First step",
                "operationId": "093c04a00ee2ab129175e1e6ca9fdda3",
                "parameters": [
                    {
                        "name": "phone_number",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/auth/client-registers/2": {
            "post": {
                "tags": [
                    "Client register"
                ],
                "summary": "last step",
                "operationId": "f016044e1c181a6cc57f7e763a5ac290",
                "parameters": [
                    {
                        "name": "client_register_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sms_code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fcm_token",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}